Move lots of things to their own components
